+++ vaadin-grid-pro.d.ts
@@ -1,2 +1,2 @@
 export * from './src/vaadin-grid-pro.js';
-export * from './@types/interfaces';
+export * from './src/interfaces';
